+++ internal/oidc/service/id_token_service_test.go
@@ -62,8 +62,9 @@ func (m *mockAccountService) ActivateAccount(_ context.Context, _ string) error 
 func (m *mockAccountService) GetAccountRoles(_ context.Context, _ string) ([]*accountDomain.Role, error) {
 	return nil, nil
 }
-func (m *mockAccountService) SetSessionRevoker(_ accountService.SessionRevoker)           {}
-func (m *mockAccountService) SetOAuth2ClientDeleter(_ accountService.OAuth2ClientDeleter) {}
+func (m *mockAccountService) SetSessionRevoker(_ accountService.SessionRevoker)                {}
+func (m *mockAccountService) SetOAuth2ClientDeleter(_ accountService.OAuth2ClientDeleter)      {}
+func (m *mockAccountService) SetConsentCacheInvalidator(_ accountService.ConsentCacheInvalidator) {}
 
 // mockCredentialRepo implements accountRepo.CredentialRepository for testing
 type mockCredentialRepo struct {